feat(react): update CTA Modal to show links
