+++ docs/app/react/cta-modal/overview.mdx
@@ -9,7 +9,7 @@ import OverviewList from '@/app/components/OverviewList'
 import DripDivider from '@/app/components/drip-divider'
 import {
   OverviewPreview,
-  WithLinkDemo,
+  WithLinksDemo,
 } from '@/app/react/cta-modal/components/cta-modal-preview'
 
 <OverviewList rules={[
@@ -25,7 +25,7 @@ import {
 
 ## With Links for actions
 
-<CodePreview preview={<WithLinkDemo />} />
+<CodePreview preview={<WithLinksDemo />} />
 
 <DripDivider />
