+++ packages/react/src/components/cta-dialog/index.ts
@@ -1,2 +1,2 @@
-export * from './context'
+export * from './provider'
 export * from './utils'
